requirements.txt
anthropic>=0.25.0
web3>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0